requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
django-crispy-forms==1.9.0
djangorestframework==3.11.0
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
xlrd==1.2.0
